Selfie detection result + callback + generic image processing service protocol

+++ oleg-verificator-example/oleg-verificator-sdk/Sources/VerificatorSDK/Services/Selfie/SelfieDetectionResponse.swift
@@ -16,5 +16,5 @@ struct SelfieDetectionResponse: Decodable {
 }
 
 struct SelfieDetectionFaceAnnotation: Decodable {
-    let detectionConfidence: Double?
+    let detectionConfidence: Float?
 }
